% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_graph}
\alias{create_graph}
\title{Create graph based on network}
\usage{
create_graph(weight_matrix)
}
\arguments{
\item{weight_matrix}{Data frame including weight matrix from a network,
infection status and gender of nodes in network.}
}
\description{
Create graph based on network
}
