% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_network}
\alias{create_network}
\title{Create initial network}
\usage{
create_network(n = 10)
}
\arguments{
\item{n}{Numeric vector of length 1. Is the number of friends in network.}
}
\value{
Data frame including weight matrix from the network,
infection status and gender of nodes (friends) in network.
}
\description{
This function creates a network based on the amount of friends a user inputs.
}
\details{
This function creates a network with user and his/her friends as nodes,
edges between user and friends, and random edges between some friends.
}
\examples{
network <- create_network(n = 15)
create_graph(network)

}
