% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_network_simulation.R
\name{sim_spread}
\alias{sim_spread}
\title{Simulate spread of disease in network}
\usage{
sim_spread(weight_matrix, spread_rate = 2)
}
\arguments{
\item{weight_matrix}{Data frame including weight matrix from a network,
infection status and gender of nodes in network.}

\item{spread_rate}{Integer that states how many nodes in network should
be infected.}
}
\description{
Simulate spread of disease in network
}
