% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNspread.R
\docType{package}
\name{SNspread}
\alias{SNspread}
\title{This package does X}
\description{
The package has several functions.
}
\section{Create Network}{

Creates initial network.
}

\section{Create Graph}{

Visualizes network in graph.
}

\section{Simulate spread}{

Simulates the spread of a disease in a network.
}

