% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{update_network}
\alias{update_network}
\title{Update network}
\usage{
update_network(weight_matrix, gender = NULL, names = NULL)
}
\arguments{
\item{gender}{Numeric vector of size number of people,
with 2 indicating women and 1 indicating men.}

\item{names}{Character vector with names of user and friends,
starting with user. Any number of names can be added,
up to the number of nodes in the network.}

\item{weigth_matrix}{Data frame including weight matrix from a network,
infection status and gender of nodes in network.}
}
\value{
Data frame including weight matrix from the network,
infection status and gender of nodes in network.
}
\description{
This function updates a network based user input.
}
\details{
This function creates a new version of a network with user and his/her
friends as nodes, edges between user and friends, and an updated name
of friends and/or addition of gender.
}
\examples{
network <- create_network(n = 15)
network2 <- update_network(network,
                           gender = c(sample(c(0,1), 16, replace = TRUE)))
create_graph(network2)

}
