% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNspread.R
\docType{package}
\name{SNspread}
\alias{SNspread}
\title{This package does creates a network of friends, plots it using a graph,
updates gender and names in network, and simulates the spread of a disease
in the network.}
\description{
The package has several functions.
}
\section{Create Network}{

Creates initial network.
}

\section{Update Network}{

Updates gender and names in network.
}

\section{Create Graph}{

Visualizes network in graph.
}

\section{Simulate Spread}{

Simulates the spread of a disease in a network.
}

